move in creating (RecordDispatcher)
I thought when we create some classes, it must move just in time.
But it was not. So, changed to move in creating of RecordDispacther.

This meen when we call stop in case of not calling start before, it will work much well. (I came up with I have not written fixes when we call stop before start.)

However, this may be educatical.